pyproject.toml
[tool.poetry]
name = "currency-averages"
version = "0.1.0"
description = ""
authors = ["Jorge Espinosa <jorge.espinosacam@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^3.141.0"
PyYAML = "^6.0"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
